+++ src/Media/ImageTo8Bit_CPP.cpp
@@ -10,18 +10,18 @@ extern "C" UOSInt ImageTo8Bit_CalDiff(Int32 col1, Int32 col2)
 	Int32 v1 = c1[0] - (Int32)c2[0];
 	UOSInt sum = 0;
 	if (v1 < 0)
-		sum += -v1;
+		sum += (UInt32)-v1;
 	else
-		sum += v1;
+		sum += (UInt32)v1;
 	v1 = c1[1] - (Int32)c2[1];
 	if (v1 < 0)
-		sum += -v1;
+		sum += (UInt32)-v1;
 	else
-		sum += v1;
+		sum += (UInt32)v1;
 	v1 = c1[2] - (Int32)c2[2];
 	if (v1 < 0)
-		sum += -v1;
+		sum += (UInt32)-v1;
 	else
-		sum += v1;
+		sum += (UInt32)v1;
 	return sum;
 }
